+++ R/package.R
@@ -46,7 +46,16 @@ keras <- NULL
     environment = "r-tensorflow",
      
     on_load = function() {
+      # check version
       check_implementation_version()
+      
+      # patch progress bar for interactive/tty sessions
+      if ((interactive() || isatty(stdout())) && keras_version() >= "2.0.9") {
+        python_path <- system.file("python", package = "keras")
+        tools <- import_from_path("kerastools", path = python_path)
+        tools$progbar$apply_patch()
+      }
+        
     },
     
     on_error = function(e) {